调整删除文件时的界面样式
将删除文件的操作放到单独的线程中(线程同步将在之后处理)
在删除文件的时候使用ProgressDialog展现进度
删除文件的进度信息由LocalBroadcastManager进行传递

+++ app/src/main/java/com/bonepeople/android/sdcardcleaner/models/SDFile.java
@@ -232,18 +232,20 @@ public class SDFile {
                     }
                 }
             } else {
-                File _file = new File(_path);
-                if (_file.delete()) {
-                    Global.set_fileCount_all(-1);
-                    Global.set_fileSize_all(-_size);
-                    if (rubbish) {
-                        Global.set_fileCount_rubbish(-1);
-                        Global.set_fileSize_rubbish(-_size);
+                if (Service_fileManager.get_deleteState() == Service_fileManager.STATE_DELETE_EXECUTING) {
+                    File _file = new File(_path);
+                    if (_file.delete()) {
+                        Global.set_fileCount_all(-1);
+                        Global.set_fileSize_all(-_size);
+                        if (rubbish) {
+                            Global.set_fileCount_rubbish(-1);
+                            Global.set_fileSize_rubbish(-_size);
+                        }
+                        if (_parent != null)
+                            _parent.updateSize(this, FILE_DELETE);
+                        _largestChild = null;
+                        _parent = null;
                     }
-                    if (_parent != null)
-                        _parent.updateSize(this, FILE_DELETE);
-                    _largestChild = null;
-                    _parent = null;
                 }
             }
         }
